+++ ELGraphics/Material_Grid.hpp
@@ -33,8 +33,8 @@ public:
 
 	virtual ~MaterialGrid() {}
 
-	int GetRowCount() const { return _rowHeights.GetSize(); }
-	int GetColumnCount() const { return _columnWidths.GetSize(); }
+	int GetRowCount() const { return (int)_rowHeights.GetSize(); }
+	int GetColumnCount() const { return (int)_columnWidths.GetSize(); }
 
 	const SharedPointer<const Texture>& GetTexture() const { return _texture; }
 	void SetDiffuse(const SharedPointer<const Texture>& texture) { _texture = texture; }
